+++ backend/canisters/user/impl/src/queries/c2c_can_issue_access_token.rs
@@ -16,6 +16,9 @@ fn c2c_can_issue_access_token_impl(args: Args, state: &RuntimeState) -> bool {
     }
 
     match args.access_type {
-        AccessTokenType::StartVideoCallV2(_) | AccessTokenType::JoinVideoCall | AccessTokenType::MarkVideoCallAsEnded => true,
+        AccessTokenType::StartVideoCallV2(_)
+        | AccessTokenType::JoinVideoCall
+        | AccessTokenType::MarkVideoCallAsEnded
+        | AccessTokenType::BotCommand(_) => true,
     }
 }